refactor: Passenger, Ticket 엔티티 및 관련 enum(vo)를 reservation 패키지 도메인 내로 통합
