+++ reservation-service/src/main/java/com/oneul_tanda/reservation_service/reservation/application/service/ReservationServiceImpl.java
@@ -3,7 +3,7 @@ package com.oneul_tanda.reservation_service.reservation.application.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.oneul_tanda.reservation_service.common.exception.CustomException;
-import com.oneul_tanda.reservation_service.passenger.domain.entity.Passenger;
+import com.oneul_tanda.reservation_service.reservation.domain.entity.Passenger;
 import com.oneul_tanda.reservation_service.reservation.application.dto.HoldReservationData;
 import com.oneul_tanda.reservation_service.reservation.application.dto.PassengerDto;
 import com.oneul_tanda.reservation_service.reservation.application.client.FlightClient;
@@ -22,8 +22,8 @@ import com.oneul_tanda.reservation_service.reservation.presentation.dto.response
 import com.oneul_tanda.reservation_service.reservation.presentation.dto.response.read.ReadReservationResponseDto;
 import com.oneul_tanda.reservation_service.reservation.presentation.dto.response.update.CancelReservationResponseDto;
 import com.oneul_tanda.reservation_service.reservation.presentation.dto.response.update.ConfirmReservationResponseDto;
-import com.oneul_tanda.reservation_service.ticket.domain.entity.SeatClass;
-import com.oneul_tanda.reservation_service.ticket.domain.entity.Ticket;
+import com.oneul_tanda.reservation_service.reservation.domain.entity.vo.SeatClass;
+import com.oneul_tanda.reservation_service.reservation.domain.entity.Ticket;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
