+++ reservation-service/src/main/java/com/oneul_tanda/reservation_service/reservation/presentation/dto/response/update/ConfirmPassengerResponseDto.java
@@ -1,7 +1,7 @@
 package com.oneul_tanda.reservation_service.reservation.presentation.dto.response.update;
 
-import com.oneul_tanda.reservation_service.passenger.domain.entity.Gender;
-import com.oneul_tanda.reservation_service.passenger.domain.entity.Passenger;
+import com.oneul_tanda.reservation_service.reservation.domain.entity.vo.Gender;
+import com.oneul_tanda.reservation_service.reservation.domain.entity.Passenger;
 
 import java.util.UUID;
